requirements.txt
Flask==3.1.1

# DB
Flask-SQLAlchemy==3.1.1
# DRIVER
mysqlclient==2.2.7

# FORMS
Flask-WTF==1.2.2

# MIGRACIONES
Flask-Migrate==4.1.0

# TESTS
pytest==8.4.1
pytest-flask==1.3.0

# ENV
python-dotenv==1.1.1